Split the frost module
